Move dependency version resolution to a properties file
With this change, we don't need to touch any Java file anymore
when updating dependency versions related to JUnit 4/5 in the Extension.
Instead, they are read from a versions.properties resource file, which
is automatically populated with the correct values from the environment
at build time. Along with this change, the plugin's apply() method has
been modularized into smaller chunks as well.