Introduce functional unit tests akin to the JUnit 5 Java plugin
This adds "FunctionalSpec" classes that verify the correct execution
of JUnit 5 tests in different scenarios. We've had some issues with
test discovery in the past, which go beyond the "pure" plugin tests
that assert the correct configuration and behavior. These new tests
create virtual projects & run their unit tests with GradleRunner,
then assert if everything went as expected.

These tests are detecting the incorrect behavior for issues like #25,
which are related to a bug in the Kotlin copy task, for instance.