Initial restructuring to support AGP3's FeaturePlugin
The ProjectConfig is now responible for determining the project type,
and throws the ProjectConfigurationException if no Android plugin is found.
Also, there's an enum describing the different kinds of projects the plugin
can work with. Since FeaturePlugin is a 3.x-only… well, feature, only
the AGP3PluginSpec will have the unit tests for this support.