Introduce DirectoryProvider interface to collect source/class dirs
Different integrations provide their own implementation of this interface
in order to be recognized by JUnit 5 tasks & companions. The current state
uses a Default (Java) implementation as well as the opt-in Kotlin provider,
applied to projects that also use the Kotlin plugin.

The Jacoco integration also profits from this new abstraction, as it
is now able to operate on _all_ source & class directories regardless
of language, which allows the respective Task to not worry about the
construction of which directories to scan anymore.