Introduce instrumentation-runner, bump versions & cut off fat (#54)
* Move AndroidJUnit5Builder into instrumentation-runner module
* Automatically apply instrumentation-test-runner library from configuration parameter
* Phase out support for Android Gradle Plugin 2, focus on latest major version
* Update to Gradle 4.5.1
* Bump minimum Gradle version to 4.3
* Change dependency scopes for instrumentation libraries, to achieve optionality

Since the instrumentation-runner library is automatically applied
by the Gradle plugin, it cannot put any JUnit 5 dependencies
on the runtime classpath itself. That would result in a premature
minSdkVersion bump for all target apps, and has to be avoided.

Instead, the instrumentation-runner only compiles against them,
and the remaining artifacts provide the APIs at runtime for only
the flavors that utilize JUnit 5 instrumentation tests.

The sample project has been enhanced to demonstrate a potential
integration of JUnit 5 instrumentation tests, using a productFlavor.