Remove Closure-based DSL & migrate to Action<T> (#123)
This type can be automatically used as a lambda expressions & Closures.
This potentially allows for a cleaner integration with both Groovy
and the Kotlin DSL for Gradle.

At the same time, the obsolete Callable0 and Callable1<T> types
are deleted now. They weren’t even used anymore, and they originally
tried to achieve the same thing. Good riddance!