Support Gradle 6 (#191)
* Reimplement reflective access in Kotlin

Gradle 6 doesn't allow removing task dependencies anymore, so it's impossible
to fix the compilation chain of a Java/Kotlin/Groovy project such that Kotlin
can call into Groovy code. Therefore, we cannot use GroovyInterop.groovy anymore.
In turn, this means that this particular class has to be rewritten in Kotlin.

Because of the stricter language, the Kotlin version looks slightly uglier than before,
but finally no Groovy code remains in the main plugin source set.

* Update to Gradle 6 & run functional tests on different Gradle versions

Test Gradle 4.7, 5.0 and the main version

* Try different memory settings on CI

* Revert "Try different memory settings on CI"

This reverts commit 7321003af5653c38bc7dcd2cfff2c145bee6be4a.

* Downgrade to Gradle 5.6.4 again

Local verification of Gradle 6 is complete, but AGP 3.2.x doesn't work with it,
and the CI's machine cannot handle multiple Gradle builds on differing versions.