Be more lenient with absent test tasks
The new Variant API, introduced in AGP 4+, allows test tasks to be
disabled on a per-variant basis with an easier API. The plugin must not
expect all test tasks to always be present, since that could result in
UnknownTaskExceptions down the line.

Now, testTaskOf() uses the less direct "findByName()" API rather than "getByName()"
to obtain the AndroidTestTasks, bailing out if a task is absent.