Introduce TaskProvider API, Remove Symlinks, Allow Connection between Plugin & Sample (#235)
* Move location of "junitPlatform" extension to Project

Don't keep it on the AGP's internal TestOptions type since that seems
to cause problems with the latest iterations of that plugin. Instead,
move it to the Project level and update all references to it

* Migrate Jacoco task integration to new lazy Gradle API

* Migrate instrumentation test integration task to new lazy TaskProvider API

* Remove symlinks

* Remove symlinks, Part 2

Some build logic is now duplicated across plugin and instrumentation,
but it's kept to a minimum. The buildSrc & deployment logic remains
in a singular place and is imported into each of the child projects
on demand

* Add changelog for Plugin 1.7.0.0

* Allow building Fat JARs of the plugin

This is in preparation for using a local build of the plugin inside
the sample modules

* Allow sample module to use Fat JAR of plugin, if possible

* Remove workaround for KotlinJvmOptions (not needed anymore)

* Changelog

* Update CI file to point to new file locations