Include runtimeOnly dependencies in generated POMs, too
This fixes a mistake in Instrumentation Core 1.2.1, which is missing
two dependencies that are essential at runtime to make the whole thing work.
The deployment script only cared about the implementation configuration
when it built the POM file for each module. Make it care about runtimeOnly as well.
(Note: compileOnly is fine to be skipped, since those dependencies are
meant to not be available at runtime)