Improve reporting for parallel instrumentation tests
Wrap the default RunNotifier with a parallel-aware variant if JUnit Jupiter's parallel test execution is enabled.
It injects itself into the default AndroidX instrumentation and reorders the emission of test events as necessary.
This requires a hefty bit of reflective inspection, therefore guard this access via a helper class.

As for the sample app, enable parallelism for it and update the tests to demonstrate the effect of it further.
Finally, improve the error message when trying to launch a UI test (e.g. Espresso) in parallel,
since that doesn't work.

Resolves #295.